+++ src/Lattice/LatticeParser.cpp
@@ -1,6 +1,8 @@
 #include <algorithm> // for std::find
+#include <utility>
 #include "LatticeParser.h"
 #include "SeriesManager.h"
+#include "SeriesParser.h"
 
 LatticeParser::LatticeParser()
 {
@@ -11,6 +13,40 @@ LatticeParser::~LatticeParser()
 {
 }
 
+void LatticeParser::addSequence(const std::string& label_in, std::string argument_in, int rank, SeriesManager *sm)
+{
+    vector <string> par;
+    string fld,val;
+    std::map <string,string> maparg;
+    maparg.insert({"label",label_in});
+    std::string seqtype;
+
+    this->chop(std::move(argument_in),&par);
+    for (int i=0;i<par.size();i++) {
+        if (par[i].empty()) {  // in the case of C: COR = {};  one is getting a zero length string.
+            continue;
+        }
+        size_t pos = par[i].find_first_of("=");
+        if (pos == string::npos) {
+            if (rank == 0) {
+                cout << "*** Warning: Ignoring invalid format: " << par[i] << " for element " << label_in << endl;
+            }
+            continue;
+        }
+        fld = par[i].substr(0, pos);
+        val = par[i].erase(0, pos + 1);
+        this->trim(fld);
+        this->trim(val);
+        if (fld.compare("type") == 0){
+            seqtype = "&sequence_"+val;
+        } else {
+            maparg.insert({fld, val});
+        }
+    }
+    auto *seriesparser = new SeriesParser;
+    seriesparser->init(rank, &maparg, seqtype, sm);
+    delete seriesparser;
+}
 
 bool LatticeParser::parse(string file, string line, int rank, vector<Element *> &lat, SeriesManager *sm)
 {
@@ -127,6 +163,17 @@ bool LatticeParser::parse(string file, string line, int rank, vector<Element *> 
   if (error){ return false; }
 
   // -------------------------------------------------------------
+  // step 2.5 - extract seuqence definition and add them to seriesmanager.
+
+  for(std::size_t i = 0; i < type.size(); i++){
+      if (type.at(i).compare("sequ") ==0) {
+          this->addSequence(label.at(i),argument.at(i),rank,sm);
+      }
+  }
+
+
+
+    // -------------------------------------------------------------
   // step 3 - resolving all references